fast resolve
az

+++ Project/Views/forum_Views.php
@@ -24,13 +24,15 @@
     </form>
 <?php
     }
-   
-    if (isset($print_message[$key]) && isset($_GET['subject_id'])) 
+    if (isset($print_message) && isset($key)) 
     {
-        echo htmlspecialchars($print_message[$key]['content_message']).' le '.
-        htmlspecialchars($print_message[$key]['date']).' par '.
-        htmlspecialchars($print_message[$key]['username']). '<a href="./index?page=forum&choix_forum=delete_message&message_num='
-        .htmlspecialchars($print_message[$key]['id']).'&subject_id='.htmlspecialchars($_SESSION['subject_id']).'"> Supprimer mon message</a><br><br>' ;
+        if (isset($print_message[$key]) && isset($_GET['subject_id'])) 
+        {
+            echo htmlspecialchars($print_message[$key]['content_message']).' le '.
+            htmlspecialchars($print_message[$key]['date']).' par '.
+            htmlspecialchars($print_message[$key]['username']). '<a href="./index?page=forum&choix_forum=delete_message&message_num='
+            .htmlspecialchars($print_message[$key]['id']).'&subject_id='.htmlspecialchars($_SESSION['subject_id']).'"> Supprimer mon message</a><br><br>' ;
+        }
     }
 
     if (isset($_GET['subject_id']) && isset($_SESSION['print_messages']) && $_SESSION['print_messages'] == false)
